+++ scripts/storage-emulator-integration/tests.ts
@@ -1,15 +1,13 @@
+import { Bucket, CopyOptions } from "@google-cloud/storage";
 import { expect } from "chai";
-import * as admin from "firebase-admin";
 import * as firebase from "firebase";
+import * as admin from "firebase-admin";
 import * as fs from "fs";
-import * as path from "path";
 import * as http from "http";
 import * as https from "https";
-import fetch from "node-fetch";
+import * as path from "path";
 import * as puppeteer from "puppeteer";
-import { Bucket, CopyOptions } from "@google-cloud/storage";
 import * as supertest from "supertest";
-
 import { IMAGE_FILE_BASE64, StorageRulesFiles } from "../../src/test/emulators/fixtures";
 import { TriggerEndToEndTest } from "../integration-helpers/framework";
 import {
@@ -23,6 +21,7 @@ import {
   readProdAppConfig,
   resetStorageEmulator,
   SERVICE_ACCOUNT_KEY,
+  signInToFirebaseAuth,
   SMALL_FILE_SIZE,
   TEST_SETUP_TIMEOUT,
   uploadText,
@@ -38,28 +37,10 @@ const TEST_CONFIG = {
   // (useful for writing tests against source of truth)
   useProductionServers: false,
 
-  // Set this to true to log all emulator logs to console
-  // (useful for debugging)
-  useMockedLogging: false,
-
   // Set this to true to make the headless chrome window visible
   // (useful for ensuring the browser is running as expected)
   showBrowser: false,
-
-  // Set this to true to keep the browser open after tests finish
-  // (useful for checking browser logs for errors)
-  keepBrowserOpen: false,
 };
-
-const EMPTY_FOLDER_DATA = `--boundary\r
-Content-Type: application/json\r
-\r
-{"contentType":"text/plain"}\r
---boundary\r
-Content-Type: text/plain\r
-\r
---boundary--\r
-`;
 
 // Temp directory to store generated files.
 let tmpDir: string;
@@ -69,7 +50,8 @@ describe("Storage emulator", () => {
   let test: TriggerEndToEndTest;
 
   let testBucket: Bucket;
-  const smallFilePath: string = createRandomFile("small_file", SMALL_FILE_SIZE, tmpDir);
+  const SMALL_FILE_PATH: string = createRandomFile("small_file", SMALL_FILE_SIZE, tmpDir);
+  const EMPTY_FILE_PATH: string = createRandomFile("empty_file", 0, tmpDir);
 
   const DEFAULT_RULES = readFile("storage.rules");
 
@@ -88,6 +70,10 @@ describe("Storage emulator", () => {
   const storageBucket = appConfig.storageBucket;
   const STORAGE_EMULATOR_HOST = getStorageEmulatorHost(emulatorConfig);
   const AUTH_EMULATOR_HOST = getAuthEmulatorHost(emulatorConfig);
+
+  const expectedHost = TEST_CONFIG.useProductionServers
+    ? "https://firebasestorage.googleapis.com"
+    : STORAGE_EMULATOR_HOST;
 
   const emulatorSpecificDescribe = TEST_CONFIG.useProductionServers ? describe.skip : describe;
 
@@ -132,10 +118,7 @@ describe("Storage emulator", () => {
     }
   });
 
-  describe("Admin SDK Endpoints", function (this) {
-    // eslint-disable-next-line @typescript-eslint/no-invalid-this
-    this.timeout(TEST_SETUP_TIMEOUT);
-
+  describe("Admin SDK Endpoints", () => {
     beforeEach(async () => {
       await resetEmulatorState();
     });
@@ -143,7 +126,7 @@ describe("Storage emulator", () => {
     describe(".bucket()", () => {
       describe("#upload()", () => {
         it("should handle non-resumable uploads", async () => {
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             resumable: false,
           });
           // Doesn't require an assertion, will throw on failure
@@ -177,7 +160,7 @@ describe("Storage emulator", () => {
         it("should handle gzip'd uploads", async () => {
           // This appears to pass, but the file gets corrupted cause it's gzipped?
           // expect(true).to.be.false;
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             gzip: true,
           });
         });
@@ -232,6 +215,34 @@ describe("Storage emulator", () => {
           });
         });
 
+        it("should handle resumable uploads with an empty buffer", async () => {
+          const fileName = "test_upload.jpg";
+          const uploadUrl = await supertest(expectedHost)
+            .post(`/v0/b/${storageBucket}/o?name=${fileName}&uploadType=resumable`)
+            .send({})
+            .set({
+              Authorization: "Bearer owner",
+              "X-Goog-Upload-Protocol": "resumable",
+              "X-Goog-Upload-Command": "start",
+            })
+            .expect(200)
+            .then((res) => {
+              return new URL(res.header["x-goog-upload-url"]);
+            });
+
+          const finalizeStatus = await supertest(expectedHost)
+            .post(uploadUrl.pathname + uploadUrl.search)
+            .send({})
+            .set({
+              Authorization: "Bearer owner",
+              "X-Goog-Upload-Protocol": "resumable",
+              "X-Goog-Upload-Command": "finalize",
+            })
+            .expect(200)
+            .then((res) => res.header["x-goog-upload-status"]);
+          expect(finalizeStatus).to.equal("final");
+        });
+
         it("should upload with provided metadata", async () => {
           const metadata = {
             contentDisposition: "attachment",
@@ -239,12 +250,50 @@ describe("Storage emulator", () => {
             contentLanguage: "de-DE",
             metadata: { foo: "bar" },
           };
-          const [, fileMetadata] = await testBucket.upload(smallFilePath, {
+          const [, fileMetadata] = await testBucket.upload(SMALL_FILE_PATH, {
             resumable: false,
             metadata,
           });
 
           expect(fileMetadata).to.deep.include(metadata);
+        });
+        // TODO(b/241151246): Fix conformance tests.
+        it("should handle resumable upload with name only in metadata", async () => {
+          const fileName = "test_upload.jpg";
+          const uploadURL = await supertest(expectedHost)
+            .post(`/upload/storage/v1/b/${storageBucket}/o?uploadType=resumable`)
+            .send({ name: fileName })
+            .set({
+              Authorization: "Bearer owner",
+            })
+            .expect(200)
+            .then((res) => new URL(res.header["location"]));
+          expect(uploadURL.searchParams?.get("name")).to.equal(fileName);
+        });
+
+        it("should handle multipart upload with name only in metadata", async () => {
+          const body = Buffer.from(`--b1d5b2e3-1845-4338-9400-6ac07ce53c1e\r
+content-type: application/json\r
+\r
+{"name":"test_upload.jpg"}\r
+--b1d5b2e3-1845-4338-9400-6ac07ce53c1e\r
+content-type: text/plain\r
+\r
+hello there!
+\r
+--b1d5b2e3-1845-4338-9400-6ac07ce53c1e--\r
+`);
+          const fileName = "test_upload.jpg";
+          const responseName = await supertest(expectedHost)
+            .post(`/upload/storage/v1/b/${storageBucket}/o?uploadType=multipart`)
+            .send(body)
+            .set({
+              Authorization: "Bearer owner",
+              "content-type": "multipart/related; boundary=b1d5b2e3-1845-4338-9400-6ac07ce53c1e",
+            })
+            .expect(200)
+            .then((res) => res.body.name);
+          expect(responseName).to.equal(fileName);
         });
 
         it("should return an error message when uploading a file with invalid metadata", async () => {
@@ -259,19 +308,19 @@ describe("Storage emulator", () => {
         });
 
         it("should be able to upload file named 'prefix/file.txt' when file named 'prefix' already exists", async () => {
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: "prefix",
           });
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: "prefix/file.txt",
           });
         });
 
         it("should be able to upload file named 'prefix' when file named 'prefix/file.txt' already exists", async () => {
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: "prefix/file.txt",
           });
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: "prefix",
           });
         });
@@ -293,7 +342,7 @@ describe("Storage emulator", () => {
               PREFIX_2_FILE,
               PREFIX_SUB_DIRECTORY_FILE,
             ].map(async (f) => {
-              await testBucket.upload(smallFilePath, {
+              await testBucket.upload(SMALL_FILE_PATH, {
                 destination: f,
               });
             })
@@ -528,7 +577,7 @@ describe("Storage emulator", () => {
           // We use a nested path to ensure that we don't need to decode
           // the objectId in the gcloud emulator API
           const bucketFilePath = "file/to/exists";
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: bucketFilePath,
           });
 
@@ -541,7 +590,7 @@ describe("Storage emulator", () => {
           // the objectId in the gcloud emulator API
           const path = "file/to";
           const bucketFilePath = path + "/exists";
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: bucketFilePath,
           });
 
@@ -555,7 +604,7 @@ describe("Storage emulator", () => {
           // We use a nested path to ensure that we don't need to decode
           // the objectId in the gcloud emulator API
           const bucketFilePath = "file/to/delete";
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination: bucketFilePath,
           });
 
@@ -585,23 +634,23 @@ describe("Storage emulator", () => {
 
       describe("#download()", () => {
         it("should return the content of the file", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [downloadContent] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .download();
 
-          const actualContent = fs.readFileSync(smallFilePath);
+          const actualContent = fs.readFileSync(SMALL_FILE_PATH);
           expect(downloadContent).to.deep.equal(actualContent);
         });
 
         it("should return partial content of the file", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [downloadContent] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             // Request 10 bytes (range requests are inclusive)
             .download({ start: 10, end: 19 });
 
-          const actualContent = fs.readFileSync(smallFilePath).slice(10, 20);
+          const actualContent = fs.readFileSync(SMALL_FILE_PATH).slice(10, 20);
           expect(downloadContent).to.have.lengthOf(10).and.deep.equal(actualContent);
         });
 
@@ -619,10 +668,12 @@ describe("Storage emulator", () => {
         const COPY_DESTINATION_FILENAME = "copied_file";
 
         it("should copy the file", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
 
           const file = testBucket.file(COPY_DESTINATION_FILENAME);
-          const [, resp] = await testBucket.file(smallFilePath.split("/").slice(-1)[0]).copy(file);
+          const [, resp] = await testBucket
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
+            .copy(file);
 
           expect(resp)
             .to.have.all.keys(["kind", "totalBytesRewritten", "objectSize", "done", "resource"])
@@ -635,33 +686,33 @@ describe("Storage emulator", () => {
 
           const [copiedContent] = await file.download();
 
-          const actualContent = fs.readFileSync(smallFilePath);
+          const actualContent = fs.readFileSync(SMALL_FILE_PATH);
           expect(copiedContent).to.deep.equal(actualContent);
         });
 
         it("should copy the file to a different bucket", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
 
           const otherBucket = testBucket.storage.bucket("other-bucket");
           const file = otherBucket.file(COPY_DESTINATION_FILENAME);
           const [, { resource: metadata }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .copy(file);
 
           expect(metadata).to.have.property("bucket", otherBucket.name);
 
           const [copiedContent] = await file.download();
 
-          const actualContent = fs.readFileSync(smallFilePath);
+          const actualContent = fs.readFileSync(SMALL_FILE_PATH);
           expect(copiedContent).to.deep.equal(actualContent);
         });
 
         it("should return the metadata of the destination file", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
 
           const file = testBucket.file(COPY_DESTINATION_FILENAME);
           const [, { resource: actualMetadata }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .copy(file);
 
           const [expectedMetadata] = await file.getMetadata();
@@ -669,7 +720,7 @@ describe("Storage emulator", () => {
         });
 
         it("should copy the file preserving the original metadata", async () => {
-          const [, source] = await testBucket.upload(smallFilePath, {
+          const [, source] = await testBucket.upload(SMALL_FILE_PATH, {
             metadata: {
               cacheControl: "private,no-store",
               metadata: {
@@ -679,7 +730,7 @@ describe("Storage emulator", () => {
           });
 
           const file = testBucket.file(COPY_DESTINATION_FILENAME);
-          await testBucket.file(smallFilePath.split("/").slice(-1)[0]).copy(file);
+          await testBucket.file(SMALL_FILE_PATH.split("/").slice(-1)[0]).copy(file);
 
           const [metadata] = await file.getMetadata();
 
@@ -725,7 +776,7 @@ describe("Storage emulator", () => {
         });
 
         it("should copy the file and overwrite with the provided custom metadata", async () => {
-          const [, source] = await testBucket.upload(smallFilePath, {
+          const [, source] = await testBucket.upload(SMALL_FILE_PATH, {
             metadata: {
               cacheControl: "private,no-store",
               metadata: {
@@ -745,7 +796,7 @@ describe("Storage emulator", () => {
             cacheControl,
           };
           const [, { resource: metadata1 }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .copy(file, copyOpts);
 
           expect(metadata1).to.deep.include({
@@ -762,7 +813,7 @@ describe("Storage emulator", () => {
         });
 
         it("should set null custom metadata values to empty strings", async () => {
-          const [, source] = await testBucket.upload(smallFilePath);
+          const [, source] = await testBucket.upload(SMALL_FILE_PATH);
 
           const file = testBucket.file(COPY_DESTINATION_FILENAME);
           const metadata = { foo: "bar", nullMetadata: null };
@@ -775,7 +826,7 @@ describe("Storage emulator", () => {
             cacheControl,
           };
           const [, { resource: metadata1 }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .copy(file, copyOpts);
 
           expect(metadata1).to.deep.include({
@@ -797,7 +848,7 @@ describe("Storage emulator", () => {
 
         it("should preserve firebaseStorageDownloadTokens", async () => {
           const firebaseStorageDownloadTokens = "token1,token2";
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             metadata: {
               metadata: {
                 firebaseStorageDownloadTokens,
@@ -807,7 +858,7 @@ describe("Storage emulator", () => {
 
           const file = testBucket.file(COPY_DESTINATION_FILENAME);
           const [, { resource: metadata }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .copy(file);
 
           expect(metadata).to.deep.include({
@@ -818,7 +869,7 @@ describe("Storage emulator", () => {
         });
 
         it("should remove firebaseStorageDownloadTokens when overwriting custom metadata", async () => {
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             metadata: {
               metadata: {
                 firebaseStorageDownloadTokens: "token1,token2",
@@ -835,18 +886,20 @@ describe("Storage emulator", () => {
             metadata,
           };
           const [, { resource: metadataOut }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .copy(file, copyOpts);
 
           expect(metadataOut).to.deep.include({ metadata });
         });
 
         it("should not support the use of a rewriteToken", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
 
           const file = testBucket.file(COPY_DESTINATION_FILENAME);
           await expect(
-            testBucket.file(smallFilePath.split("/").slice(-1)[0]).copy(file, { token: "foo-bar" })
+            testBucket
+              .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
+              .copy(file, { token: "foo-bar" })
           ).to.eventually.be.rejected.and.have.property("code", 501);
         });
       });
@@ -854,7 +907,7 @@ describe("Storage emulator", () => {
       describe("#makePublic()", () => {
         it("should no-op", async () => {
           const destination = "a/b";
-          await testBucket.upload(smallFilePath, { destination });
+          await testBucket.upload(SMALL_FILE_PATH, { destination });
           const [aclMetadata] = await testBucket.file(destination).makePublic();
 
           const generation = aclMetadata.generation;
@@ -876,7 +929,7 @@ describe("Storage emulator", () => {
 
         it("should not interfere with downloading of bytes via public URL", async () => {
           const destination = "a/b";
-          await testBucket.upload(smallFilePath, { destination });
+          await testBucket.upload(SMALL_FILE_PATH, { destination });
           await testBucket.file(destination).makePublic();
 
           const publicLink = `${STORAGE_EMULATOR_HOST}/${testBucket.name}/${destination}`;
@@ -901,7 +954,7 @@ describe("Storage emulator", () => {
         it("should throw on non-existing file", async () => {
           let err: any;
           await testBucket
-            .file(smallFilePath)
+            .file(SMALL_FILE_PATH)
             .getMetadata()
             .catch((_err) => {
               err = _err;
@@ -911,9 +964,9 @@ describe("Storage emulator", () => {
         });
 
         it("should return generated metadata for new upload", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .getMetadata();
 
           const metadataTypes: { [s: string]: string } = {};
@@ -979,9 +1032,9 @@ describe("Storage emulator", () => {
         });
 
         it("should return a functional media link", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [{ mediaLink }] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .getMetadata();
 
           const requestClient = TEST_CONFIG.useProductionServers ? https : http;
@@ -1001,7 +1054,7 @@ describe("Storage emulator", () => {
 
         it("should handle firebaseStorageDownloadTokens", async () => {
           const destination = "public/small_file";
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination,
             metadata: {},
           });
@@ -1047,7 +1100,7 @@ describe("Storage emulator", () => {
         it("should throw on non-existing file", async () => {
           let err: any;
           await testBucket
-            .file(smallFilePath)
+            .file(SMALL_FILE_PATH)
             .setMetadata({ contentType: 9000 })
             .catch((_err) => {
               err = _err;
@@ -1057,9 +1110,9 @@ describe("Storage emulator", () => {
         });
 
         it("should allow overriding of default metadata", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ contentType: "very/fake" });
 
           const metadataTypes: { [s: string]: string } = {};
@@ -1096,9 +1149,9 @@ describe("Storage emulator", () => {
         });
 
         it("should allow setting of optional metadata", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ cacheControl: "no-cache", contentLanguage: "en" });
 
           const metadataTypes: { [s: string]: string } = {};
@@ -1115,7 +1168,7 @@ describe("Storage emulator", () => {
 
         it("should not duplicate data when called repeatedly", async () => {
           const destination = "public/small_file";
-          await testBucket.upload(smallFilePath, {
+          await testBucket.upload(SMALL_FILE_PATH, {
             destination,
             metadata: {},
           });
@@ -1151,18 +1204,18 @@ describe("Storage emulator", () => {
         });
 
         it("should allow fields under .metadata", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ metadata: { is_over: "9000" } });
 
           expect(metadata.metadata.is_over).to.equal("9000");
         });
 
         it("should convert non-string fields under .metadata to strings", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ metadata: { booleanValue: true, numberValue: -1 } });
 
           expect(metadata.metadata).to.deep.equal({
@@ -1172,9 +1225,9 @@ describe("Storage emulator", () => {
         });
 
         it("should remove fields under .metadata when setting to null", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata1] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ metadata: { foo: "bar", hello: "world" } });
 
           expect(metadata1.metadata).to.deep.equal({
@@ -1183,7 +1236,7 @@ describe("Storage emulator", () => {
           });
 
           const [metadata2] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ metadata: { foo: null } });
 
           expect(metadata2.metadata).to.deep.equal({
@@ -1192,9 +1245,9 @@ describe("Storage emulator", () => {
         });
 
         it("should ignore any unknown fields", async () => {
-          await testBucket.upload(smallFilePath);
+          await testBucket.upload(SMALL_FILE_PATH);
           const [metadata] = await testBucket
-            .file(smallFilePath.split("/").slice(-1)[0])
+            .file(SMALL_FILE_PATH.split("/").slice(-1)[0])
             .setMetadata({ nada: "true" });
 
           expect(metadata.nada).to.be.undefined;
@@ -1223,11 +1276,6 @@ describe("Storage emulator", () => {
         headless: !TEST_CONFIG.showBrowser,
         devtools: true,
       });
-    });
-
-    beforeEach(async function (this) {
-      this.timeout(TEST_SETUP_TIMEOUT);
-
       page = await browser.newPage();
       await page.goto("https://example.com", { waitUntil: "networkidle2" });
 
@@ -1255,25 +1303,28 @@ describe("Storage emulator", () => {
         AUTH_EMULATOR_HOST,
         STORAGE_EMULATOR_HOST.replace(/^(https?:|)\/\//, "")
       );
+    });
 
+    beforeEach(async () => {
       await resetEmulatorState();
-
       await testBucket.upload(image_filename, { destination: filename });
     });
 
     afterEach(async () => {
-      await page.close();
+      await page.evaluate(async () => {
+        await firebase.auth().signOut();
+      });
     });
 
     after(async () => {
+      await page.close();
       await browser.close();
     });
 
     describe(".ref()", () => {
       describe("#put()", () => {
-        it("should upload a file", async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
-
+        it("should upload a file", async () => {
+          await signInToFirebaseAuth(page);
           const uploadState = await uploadText(
             page,
             "testing/image.png",
@@ -1284,9 +1335,8 @@ describe("Storage emulator", () => {
           expect(uploadState).to.equal("success");
         });
 
-        it("should upload a file with a really long path name to check for os filename character limit", async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
-
+        it("should upload a file with a really long path name to check for os filename character limit", async () => {
+          await signInToFirebaseAuth(page);
           const uploadState = await uploadText(
             page,
             `testing/${"long".repeat(180)}image.png`,
@@ -1297,8 +1347,7 @@ describe("Storage emulator", () => {
           expect(uploadState).to.equal("success");
         });
 
-        it("should upload replace existing file", async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
+        it("should upload replace existing file", async () => {
           await uploadText(page, "upload/replace.txt", "some-content");
           await uploadText(page, "upload/replace.txt", "some-other-content");
 
@@ -1332,6 +1381,7 @@ describe("Storage emulator", () => {
         });
 
         it("should upload a file using put", async () => {
+          await signInToFirebaseAuth(page);
           const uploadState = await page.evaluate(async (IMAGE_FILE_BASE64) => {
             const task = await firebase
               .storage()
@@ -1339,6 +1389,16 @@ describe("Storage emulator", () => {
               .put(new File([IMAGE_FILE_BASE64], "toUpload.txt"));
             return task.state;
           }, IMAGE_FILE_BASE64);
+
+          expect(uploadState).to.equal("success");
+        });
+
+        it("should handle uploading empty buffer", async () => {
+          await signInToFirebaseAuth(page);
+          const uploadState = await page.evaluate(async () => {
+            const task = await firebase.storage().ref("testing/empty_file").put(new ArrayBuffer(0));
+            return task.state;
+          });
 
           expect(uploadState).to.equal("success");
         });
@@ -1406,241 +1466,119 @@ describe("Storage emulator", () => {
           }, IMAGE_FILE_BASE64);
           expect(uploadState!).to.include("User does not have permission");
         });
+
+        it("should return a 403 on rules deny when overwriting existing file", async () => {
+          async function shouldThrowOnUpload() {
+            try {
+              return await uploadText(
+                page,
+                "upload/allowIfNoExistingFile.txt",
+                "some-other-content"
+              );
+            } catch (err: any) {
+              if (err instanceof Error) {
+                return err.message;
+              }
+              throw err;
+            }
+          }
+
+          await uploadText(page, "upload/allowIfNoExistingFile.txt", "some-content");
+
+          const uploadState = await shouldThrowOnUpload();
+          expect(uploadState!).to.include("User does not have permission");
+        });
       });
 
       describe("#listAll()", () => {
-        beforeEach(async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
+        async function uploadFiles(paths: string[], filename = SMALL_FILE_PATH): Promise<void> {
+          await Promise.all(
+            paths.map((destination) => testBucket.upload(filename, { destination }))
+          );
+        }
 
-          const refs = [
-            "testing/storage_ref/image.png",
-            "testing/somePathEndsWithDoubleSlash//file.png",
-          ];
-          for (const ref of refs) {
-            await testBucket.upload(smallFilePath, { destination: ref });
-          }
-        });
-
-        // TODO(b/240637118): Skipping due to listAll functionality being broken.
-        it.skip("should list all files and prefixes", async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
-
-          const refs = [
-            "listAll/subdir/storage_ref/image.png",
-            "listAll/subdir/somePathEndsWithDoubleSlash//file.png",
-            "listAll/subdir/item1",
-            "listAll/subdir/item2",
-          ];
-          for (const ref of refs) {
-            await testBucket.upload(smallFilePath, { destination: ref });
-          }
-
-          const listResult = await page.evaluate(async () => {
-            const list = await firebase.storage().ref("listAll/subdir/").listAll();
+        async function executeListAllAtPath(path: string): Promise<{
+          items: string[];
+          prefixes: string[];
+        }> {
+          return await page.evaluate(async (path) => {
+            const list = await firebase.storage().ref(path).listAll();
             return {
               prefixes: list.prefixes.map((prefix) => prefix.name),
               items: list.items.map((item) => item.name),
             };
-          });
+          }, path);
+        }
+
+        it("should list all files and prefixes at path", async () => {
+          await uploadFiles([
+            "listAll/some/deeply/nested/directory/item1",
+            "listAll/item1",
+            "listAll/item2",
+          ]);
+
+          const listResult = await executeListAllAtPath("listAll/");
 
           expect(listResult).to.deep.equal({
             items: ["item1", "item2"],
-            prefixes: ["somePathEndsWithDoubleSlash", "storage_ref"],
-          });
-        });
-
-        it("should list implicit prefixes", async () => {
-          await page.evaluate(
-            async (IMAGE_FILE_BASE64, filename) => {
-              try {
-                await firebase.auth().signInAnonymously();
-                const task = await firebase
-                  .storage()
-                  .ref(filename)
-                  .putString(IMAGE_FILE_BASE64, "base64");
-                return task.state;
-              } catch (err: any) {
-                throw err.message;
-              }
-            },
-            IMAGE_FILE_BASE64,
-            `testing/implicit/deep/path/file.jpg`
-          );
-
-          const listResult = await page.evaluate(async () => {
-            const list = await firebase.storage().ref("testing/implicit").listAll();
-            return {
-              prefixes: list.prefixes.map((prefix) => prefix.name),
-              items: list.items.map((item) => item.name),
-            };
-          });
-
-          expect(listResult).to.deep.equal({
-            prefixes: ["deep"],
-            items: [],
-          });
-        });
-
-        it("should list at /", async () => {
-          await uploadText(page, "list/file.jpg", "hello");
-          await uploadText(page, "list/subdir/file.jpg", "world");
-
-          const listResult = await page.evaluate(async () => {
-            const list = await firebase.storage().ref("/list").listAll();
-            return {
-              prefixes: list.prefixes.map((prefix) => prefix.name),
-              items: list.items.map((item) => item.name),
-            };
-          });
-
-          expect(listResult).to.deep.equal({
-            prefixes: ["subdir"],
-            items: ["file.jpg"],
+            prefixes: ["some"],
           });
         });
 
         it("zero element list array should still be present in response", async () => {
-          const listResult = await page.evaluate(async () => {
-            const list = await firebase.storage().ref("/list").listAll();
-            return {
-              prefixes: list.prefixes.map((prefix) => prefix.name),
-              items: list.items.map((item) => item.name),
-            };
-          });
+          const listResult = await executeListAllAtPath("listAll/");
 
           expect(listResult).to.deep.equal({
             prefixes: [],
             items: [],
           });
         });
-        context("with folder placeholders", () => {
-          beforeEach(async function (this) {
-            this.timeout(TEST_SETUP_TIMEOUT);
 
-            const refs = [
-              "testing/abc", // empty folder inside testing/
-              "testing/storage_ref", // also an implicit prefix with files
-            ];
-            for (const ref of refs) {
-              // Use REST API to create the folder placeholders since SDK won't
-              // allow refs with trailing slashes.
-              await fetch(
-                `${STORAGE_EMULATOR_HOST}/upload/storage/v1/b/${storageBucket}/o?name=${encodeURIComponent(
-                  ref
-                )}/`,
-                {
-                  headers: {
-                    "Content-Type": "multipart/related; boundary=boundary",
-                  },
-                  method: "POST",
-                  body: Buffer.from(EMPTY_FOLDER_DATA, "utf8"),
-                }
-              );
-            }
+        it("folder placeholder should not be listed under itself", async () => {
+          await uploadFiles(["listAll/abc/", EMPTY_FILE_PATH]);
+
+          let listResult = await executeListAllAtPath("listAll/");
+
+          expect(listResult).to.deep.equal({
+            prefixes: ["abc"],
+            items: [],
           });
 
-          it("folder placeholder should not be listed under itself", async () => {
-            const listResult = await page.evaluate(async () => {
-              const list = await firebase.storage().ref("/testing/abc").listAll();
-              return {
-                prefixes: list.prefixes.map((prefix) => prefix.name),
-                items: list.items.map((item) => item.name),
-              };
-            });
+          listResult = await executeListAllAtPath("listAll/abc/");
 
-            expect(listResult).to.deep.equal({
-              prefixes: [],
-              items: [],
-            });
-          });
-
-          it("folder placeholder should be listed as a prefix but not an item under parent", async () => {
-            const listResult = await page.evaluate(async () => {
-              const list = await firebase.storage().ref("/testing").listAll();
-              return {
-                prefixes: list.prefixes.map((prefix) => prefix.name),
-                items: list.items.map((item) => item.name),
-              };
-            });
-
-            expect(listResult).to.deep.equal({
-              prefixes: ["abc", "somePathEndsWithDoubleSlash", "storage_ref"],
-              items: [],
-            });
+          expect(listResult).to.deep.equal({
+            prefixes: [],
+            items: [],
           });
         });
 
-        context("with invalid prefixes and items", () => {
-          beforeEach(async function (this) {
-            this.timeout(TEST_SETUP_TIMEOUT);
+        it("should not include show invalid prefixes and items", async () => {
+          await uploadFiles(["listAll//foo", "listAll/bar//", "listAll/baz//qux"], EMPTY_FILE_PATH);
 
-            const refs = ["list//foo", "list/bar//", "list/baz//qux"];
-            for (const ref of refs) {
-              // Use REST API to create the folder placeholders since SDK won't
-              // allow refs with trailing slashes.
-              await fetch(
-                `${STORAGE_EMULATOR_HOST}/upload/storage/v1/b/${storageBucket}/o?name=${encodeURIComponent(
-                  ref
-                )}`,
-                {
-                  headers: {
-                    "Content-Type": "multipart/related; boundary=boundary",
-                  },
-                  method: "POST",
-                  body: Buffer.from(EMPTY_FOLDER_DATA, "utf8"),
-                }
-              );
-            }
-          });
+          const listResult = await executeListAllAtPath("listAll/");
 
-          it("list result should not include show invalid prefixes and items", async () => {
-            const listResult = await page.evaluate(async () => {
-              const list = await firebase.storage().ref("/list").listAll();
-              return {
-                prefixes: list.prefixes.map((prefix) => prefix.name),
-                items: list.items.map((item) => item.name),
-              };
-            });
-
-            expect(listResult).to.deep.equal({
-              prefixes: ["bar", "baz"], // only implicit prefixes, (no bar//)
-              items: [], // no valid items
-            });
+          expect(listResult).to.deep.equal({
+            prefixes: ["bar", "baz"],
+            items: [], // no valid items
           });
         });
       });
 
       describe("#list()", () => {
+        async function uploadFiles(paths: string[]): Promise<void> {
+          await Promise.all(
+            paths.map((destination) => testBucket.upload(SMALL_FILE_PATH, { destination }))
+          );
+        }
         const itemNames = [...Array(10)].map((_, i) => `item#${i}`);
 
-        beforeEach(async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
-
-          for (const item of itemNames) {
-            await page.evaluate(
-              async (IMAGE_FILE_BASE64, filename) => {
-                try {
-                  await firebase.auth().signInAnonymously();
-                  const task = await firebase
-                    .storage()
-                    .ref(filename)
-                    .putString(IMAGE_FILE_BASE64, "base64");
-                  return task.state;
-                } catch (err: any) {
-                  throw err.message;
-                }
-              },
-              IMAGE_FILE_BASE64,
-              `testing/list/${item}`
-            );
-          }
+        beforeEach(async () => {
+          await uploadFiles(itemNames.map((name) => `listAll/${name}`));
         });
 
-        it("should list only maxResults items with nextPageToken, when maxResults is set", async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
-
+        it("should list only maxResults items with nextPageToken, when maxResults is set", async () => {
           const listItems = await page.evaluate(async () => {
-            const list = await firebase.storage().ref("testing/list").list({
+            const list = await firebase.storage().ref("listAll").list({
               maxResults: 4,
             });
             return {
@@ -1654,15 +1592,14 @@ describe("Storage emulator", () => {
           expect(listItems.nextPageToken).to.not.be.empty;
         });
 
-        it("should paginate when nextPageToken is provided", async function (this) {
-          this.timeout(TEST_SETUP_TIMEOUT);
+        it("should paginate when nextPageToken is provided", async () => {
           let responses: string[] = [];
           let pageToken = "";
           let pageCount = 0;
 
           do {
             const listResponse = await page.evaluate(async (pageToken) => {
-              const list = await firebase.storage().ref("testing/list").list({
+              const list = await firebase.storage().ref("listAll").list({
                 maxResults: 4,
                 pageToken,
               });
@@ -1685,59 +1622,19 @@ describe("Storage emulator", () => {
         });
       });
 
-      it("updateMetadata throws on non-existent file", async () => {
-        const err = await page.evaluate(async () => {
-          try {
-            return await firebase
-              .storage()
-              .ref("testing/thisFileDoesntExist")
-              .updateMetadata({
-                contentType: "application/awesome-stream",
-                customMetadata: {
-                  testable: "true",
-                },
-              });
-          } catch (_err) {
-            return _err;
-          }
-        });
-
-        expect(err).to.not.be.empty;
-      });
-
-      it("updateMetadata updates metadata successfully", async () => {
-        const metadata = await page.evaluate(async (filename) => {
-          await firebase.auth().signInAnonymously();
-          return firebase
-            .storage()
-            .ref(filename)
-            .updateMetadata({
-              contentType: "application/awesome-stream",
-              customMetadata: {
-                testable: "true",
-              },
-            });
-        }, filename);
-
-        expect(metadata.contentType).to.equal("application/awesome-stream");
-        expect(metadata.customMetadata.testable).to.equal("true");
-      });
-
       describe("#getDownloadURL()", () => {
         it("returns url pointing to the expected host", async () => {
+          await signInToFirebaseAuth(page);
           const downloadUrl: string = await page.evaluate((filename) => {
             return firebase.storage().ref(filename).getDownloadURL();
           }, filename);
-          const expectedHost = TEST_CONFIG.useProductionServers
-            ? "https://firebasestorage.googleapis.com"
-            : STORAGE_EMULATOR_HOST;
-
           expect(downloadUrl).to.contain(
             `${expectedHost}/v0/b/${storageBucket}/o/testing%2Fstorage_ref%2Fimage.png?alt=media&token=`
           );
         });
 
         it("serves the right content", async () => {
+          await signInToFirebaseAuth(page);
           const downloadUrl = await page.evaluate((filename) => {
             return firebase.storage().ref(filename).getDownloadURL();
           }, filename);
@@ -1760,83 +1657,115 @@ describe("Storage emulator", () => {
         });
       });
 
-      it("#getMetadata()", async () => {
-        const metadata = await page.evaluate((filename) => {
-          return firebase.storage().ref(filename).getMetadata();
-        }, filename);
-
-        const metadataTypes: { [s: string]: string } = {};
-
-        for (const key in metadata) {
-          if (metadata[key]) {
-            metadataTypes[key] = typeof metadata[key];
-          }
-        }
-
-        expect(metadataTypes).to.deep.equal({
-          bucket: "string",
-          contentDisposition: "string",
-          contentEncoding: "string",
-          contentType: "string",
-          cacheControl: "string",
-          fullPath: "string",
-          generation: "string",
-          md5Hash: "string",
-          metageneration: "string",
-          name: "string",
-          size: "number",
-          timeCreated: "string",
-          type: "string",
-          updated: "string",
-        });
-      });
-
-      describe("#setMetadata()", () => {
-        it("should allow for custom metadata to be set", async () => {
+      describe("#getMetadata()", () => {
+        it("should return file metadata", async () => {
+          await signInToFirebaseAuth(page);
           const metadata = await page.evaluate(async (filename) => {
-            await firebase
-              .storage()
-              .ref(filename)
-              .updateMetadata({
-                customMetadata: {
-                  is_over: "9000",
-                },
-              });
             return await firebase.storage().ref(filename).getMetadata();
           }, filename);
 
-          expect(metadata.customMetadata.is_over).to.equal("9000");
+          const metadataTypes: { [s: string]: string } = {};
+
+          for (const key in metadata) {
+            if (metadata[key]) {
+              metadataTypes[key] = typeof metadata[key];
+            }
+          }
+
+          expect(metadataTypes).to.deep.equal({
+            bucket: "string",
+            contentDisposition: "string",
+            contentEncoding: "string",
+            contentType: "string",
+            cacheControl: "string",
+            fullPath: "string",
+            generation: "string",
+            md5Hash: "string",
+            metageneration: "string",
+            name: "string",
+            size: "number",
+            timeCreated: "string",
+            type: "string",
+            updated: "string",
+          });
+        });
+      });
+
+      describe("#updateMetadata()", () => {
+        it("updates metadata successfully", async () => {
+          await signInToFirebaseAuth(page);
+          const metadata = await page.evaluate(async (filename) => {
+            return firebase
+              .storage()
+              .ref(filename)
+              .updateMetadata({
+                contentType: "application/awesome-stream",
+                customMetadata: {
+                  testable: "true",
+                },
+              });
+          }, filename);
+
+          expect(metadata.contentType).to.equal("application/awesome-stream");
+          expect(metadata.customMetadata.testable).to.equal("true");
         });
 
         it("should allow deletion of custom metadata by setting to null", async () => {
+          await signInToFirebaseAuth(page);
           const setMetadata = await page.evaluate((filename) => {
-            const storageReference = firebase.storage().ref(filename);
-            return storageReference.updateMetadata({
-              contentType: "text/plain",
-              customMetadata: {
-                removeMe: "please",
-              },
-            });
+            return firebase
+              .storage()
+              .ref(filename)
+              .updateMetadata({
+                contentType: "text/plain",
+                customMetadata: {
+                  removeMe: "please",
+                },
+              });
           }, filename);
 
           expect(setMetadata.customMetadata.removeMe).to.equal("please");
 
           const nulledMetadata = await page.evaluate((filename) => {
-            const storageReference = firebase.storage().ref(filename);
-            return storageReference.updateMetadata({
-              contentType: "text/plain",
-              customMetadata: {
-                removeMe: null as any,
-              },
-            });
+            return firebase
+              .storage()
+              .ref(filename)
+              .updateMetadata({
+                contentType: "text/plain",
+                customMetadata: {
+                  removeMe: null as any,
+                },
+              });
           }, filename);
 
           expect(nulledMetadata.customMetadata.removeMe).to.equal(undefined);
+        });
+
+        it("throws on non-existent file", async () => {
+          await signInToFirebaseAuth(page);
+          const err = await page.evaluate(async () => {
+            try {
+              return await firebase
+                .storage()
+                .ref("testing/thisFileDoesntExist")
+                .updateMetadata({
+                  contentType: "application/awesome-stream",
+                  customMetadata: {
+                    testable: "true",
+                  },
+                });
+            } catch (_err) {
+              return _err;
+            }
+          });
+
+          expect(err).to.not.be.empty;
         });
       });
 
       describe("deleteFile", () => {
         it("should delete file", async () => {
+          await signInToFirebaseAuth(page);
           await page.evaluate((filename) => {
             return firebase.storage().ref(filename).delete();
           }, filename);
@@ -1888,8 +1817,7 @@ describe("Storage emulator", () => {
       tmpDir
     );
 
-    beforeEach(async function (this) {
-      this.timeout(TEST_SETUP_TIMEOUT);
+    beforeEach(async () => {
       await resetEmulatorState();
       await testBucket.upload(image_filename, { destination: filename });
     });