+++ src/emulator/types.ts
@@ -13,6 +13,7 @@ export enum Emulators {
   LOGGING = "logging",
   STORAGE = "storage",
   EXTENSIONS = "extensions",
+  EVENTARC = "eventarc",
 }
 
 export type DownloadableEmulators =
@@ -45,6 +46,7 @@ export const ALL_SERVICE_EMULATORS = [
   Emulators.HOSTING,
   Emulators.PUBSUB,
   Emulators.STORAGE,
+  Emulators.EVENTARC,
 ].filter((v) => v);
 
 export const EMULATORS_SUPPORTED_BY_FUNCTIONS = [
@@ -52,6 +54,7 @@ export const EMULATORS_SUPPORTED_BY_FUNCTIONS = [
   Emulators.DATABASE,
   Emulators.PUBSUB,
   Emulators.STORAGE,
+  Emulators.EVENTARC,
 ];
 
 export const EMULATORS_SUPPORTED_BY_UI = [
@@ -68,6 +71,7 @@ export const EMULATORS_SUPPORTED_BY_USE_EMULATOR = [
   Emulators.DATABASE,
   Emulators.FIRESTORE,
   Emulators.FUNCTIONS,
+  Emulators.STORAGE,
 ];
 
 // TODO: Is there a way we can just allow iteration over the enum?