+++ src/test/emulators/storage/files.spec.ts
@@ -47,12 +47,32 @@ describe("files", () => {
     let _persistence: Persistence;
     let _uploadService: UploadService;
 
+    type UploadFileOptions = {
+      data?: string;
+      metadata?: Object;
+    };
+
+    async function uploadFile(
+      storageLayer: StorageLayer,
+      bucketId: string,
+      objectId: string,
+      opts?: UploadFileOptions
+    ) {
+      const upload = _uploadService.multipartUpload({
+        bucketId,
+        objectId: encodeURIComponent(objectId),
+        dataRaw: Buffer.from(opts?.data ?? "hello world"),
+        metadataRaw: JSON.stringify(opts?.metadata ?? {}),
+      });
+      await storageLayer.uploadObject(upload);
+    }
+
     beforeEach(() => {
       _persistence = new Persistence(getPersistenceTmpDir());
       _uploadService = new UploadService(_persistence);
     });
 
-    describe("#handleUploadObject()", () => {
+    describe("#uploadObject()", () => {
       it("should throw if upload is not finished", () => {
         const storageLayer = getStorageLayer(ALWAYS_TRUE_RULES_VALIDATOR);
         const upload = _uploadService.startResumableUpload({
@@ -78,16 +98,13 @@ describe("files", () => {
       });
     });
 
-    describe("#handleGetObject()", () => {
+    describe("#getObject()", () => {
       it("should return data and metadata", async () => {
         const storageLayer = getStorageLayer(ALWAYS_TRUE_RULES_VALIDATOR);
-        const upload = _uploadService.multipartUpload({
-          bucketId: "bucket",
-          objectId: "dir%2Fobject",
-          metadataRaw: `{"contentType": "mime/type"}`,
-          dataRaw: Buffer.from("Hello, World!"),
+        await uploadFile(storageLayer, "bucket", "dir/object", {
+          data: "Hello, World!",
+          metadata: { contentType: "mime/type" },
         });
-        await storageLayer.uploadObject(upload);
 
         const { metadata, data } = await storageLayer.getObject({
           bucketId: "bucket",